Fetch latest contact voice message via existing route
